researchAgent/requirements.txt
crewai
crewai[tools]
langchain_community
langchain_groq
python-dotenv
